+++ libs/ymir-core/include/ymir/hw/scu/scu.hpp
@@ -117,7 +117,7 @@ public:
     // External interrupt triggers
 
     void UpdateHBlank(bool hb, bool vb);
-    void UpdateVBlank(bool hb, bool vb);
+    void UpdateVBlank(bool vb);
     void TriggerTimer0();
     void TriggerTimer1();
     void TriggerDSPEnd();
@@ -130,7 +130,6 @@ public:
 
     void TriggerExternalInterrupt0();
     void AcknowledgeExternalInterrupt();
-    void AcknowledgeExternalSlaveInterrupt();
 
     // -------------------------------------------------------------------------
     // RAM/register dumps
@@ -182,7 +181,6 @@ private:
     uint8 m_pendingIntrIndex;
 
     void UpdateMasterInterruptLevel();
-    void UpdateSlaveInterruptLevel(bool hb, bool vb);
 
     // -------------------------------------------------------------------------
     // DMA
@@ -278,9 +276,9 @@ public:
     const sh2::CBAcknowledgeExternalInterrupt CbAckExtIntr =
         util::MakeClassMemberRequiredCallback<&SCU::AcknowledgeExternalInterrupt>(this);
 
-    const vdp::CBHVBlankStateChange CbHBlankStateChange =
+    const vdp::CBHBlankStateChange CbHBlankStateChange =
         util::MakeClassMemberRequiredCallback<&SCU::UpdateHBlank>(this);
-    const vdp::CBHVBlankStateChange CbVBlankStateChange =
+    const vdp::CBVBlankStateChange CbVBlankStateChange =
         util::MakeClassMemberRequiredCallback<&SCU::UpdateVBlank>(this);
     const vdp::CBTriggerEvent CbTriggerSpriteDrawEnd =
         util::MakeClassMemberRequiredCallback<&SCU::TriggerSpriteDrawEnd>(this);