+++ libs/ymir-core/include/ymir/hw/vdp/vdp.hpp
@@ -42,7 +42,7 @@ public:
 
     void Reset(bool hard);
 
-    void MapCallbacks(CBHVBlankStateChange cbHBlankStateChange, CBHVBlankStateChange cbVBlankStateChange,
+    void MapCallbacks(CBHBlankStateChange cbHBlankStateChange, CBVBlankStateChange cbVBlankStateChange,
                       CBTriggerEvent cbSpriteDrawEnd, CBTriggerEvent cbOptimizedINTBACKRead) {
         m_cbHBlankStateChange = cbHBlankStateChange;
         m_cbVBlankStateChange = cbVBlankStateChange;
@@ -119,8 +119,8 @@ private:
     // Only valid when color RAM mode is one of the RGB555 modes.
     alignas(16) std::array<Color888, kVDP2CRAMSize / sizeof(uint16)> m_CRAMCache;
 
-    CBHVBlankStateChange m_cbHBlankStateChange;
-    CBHVBlankStateChange m_cbVBlankStateChange;
+    CBHBlankStateChange m_cbHBlankStateChange;
+    CBVBlankStateChange m_cbVBlankStateChange;
     CBTriggerEvent m_cbTriggerSpriteDrawEnd;
     CBTriggerEvent m_cbTriggerOptimizedINTBACKRead;