+++ libs/ymir-core/include/ymir/hw/vdp/vdp_internal_callbacks.hpp
@@ -9,8 +9,11 @@
 
 namespace ymir::vdp {
 
-/// @brief Invoked when the HBlank and VBlank signals change.
-using CBHVBlankStateChange = util::RequiredCallback<void(bool hb, bool vb)>;
+/// @brief Invoked when the HBlank signal changes.
+using CBHBlankStateChange = util::RequiredCallback<void(bool hb, bool vb)>;
+
+/// @brief Invoked when the VBlank signal changes.
+using CBVBlankStateChange = util::RequiredCallback<void(bool vb)>;
 
 /// @brief Invoked when specific events occur while processing.
 using CBTriggerEvent = util::RequiredCallback<void()>;